Split SS_Shipping_WC_Order into single-responsibility components
class-ss-shipping-wc-order.php (1432 lines) is split along its
responsibility lines into:

- SS_Shipping_Order_Meta (416): order meta access - resolved method id,
  pick-up point agent meta (incl. the update/delete meta hooks and the
  API-validating save_shipping_agent), parcel split, shipment ids.
- SS_Shipping_Order_Meta_Box (280): the order screen meta box
  (registration + rendering + weight/address formatting).
- SS_Shipping_Label_Creator (435): the AJAX endpoint and the full label
  creation flow (shipment API call, order note, tracking, PDF saving,
  label links).
- SS_Shipping_Order_Bulk_Actions (373): bulk label generation on the
  Orders screen incl. combined-PDF creation.
- SS_Shipping_WC_Order (364): facade wiring hooks and delegating; its
  full public surface (and the protected methods the characterization
  tests invoke via reflection: save_shipping_agent,
  create_label_for_single_order_maybe_return) is unchanged, so
  SS_SHIPPING_WC()->get_ss_shipping_wc_order() callers and merchant
  snippets keep working. All hooks are still registered with facade
  callables, so third-party remove_filter/remove_action calls are
  unaffected. Code bodies are moved verbatim.

Integration suite green (150 passed), phpcs clean.
